Move HAPI test to its own file Construct fixture to handle mocking resource_show and resource_patch Update dependency versions Update pre-commit
