+++ requirements.txt
@@ -16,19 +16,19 @@ cfgv==3.4.0
     # via pre-commit
 chardet==5.2.0
     # via frictionless
-charset-normalizer==3.3.2
+charset-normalizer==3.4.0
     # via requests
 ckanapi==4.8
     # via hdx-python-api (pyproject.toml)
 click==8.1.7
     # via typer
-coverage==7.6.1
+coverage==7.6.4
     # via pytest-cov
 defopt==6.4.0
     # via hdx-python-api (pyproject.toml)
-distlib==0.3.8
+distlib==0.3.9
     # via virtualenv
-dnspython==2.6.1
+dnspython==2.7.0
     # via email-validator
 docopt==0.6.2
     # via
@@ -38,31 +38,31 @@ docutils==0.21.2
     # via defopt
 email-validator==2.2.0
     # via hdx-python-api (pyproject.toml)
-et-xmlfile==1.1.0
+et-xmlfile==2.0.0
     # via openpyxl
-filelock==3.16.0
+filelock==3.16.1
     # via virtualenv
-frictionless==5.17.1
+frictionless==5.18.0
     # via hdx-python-utilities
-google-auth==2.34.0
+google-auth==2.36.0
     # via
     #   google-auth-oauthlib
     #   gspread
 google-auth-oauthlib==1.2.1
     # via gspread
-gspread==6.1.2
+gspread==6.1.4
     # via hdx-python-api (pyproject.toml)
-hdx-python-country==3.7.8
+hdx-python-country==3.8.3
     # via hdx-python-api (pyproject.toml)
-hdx-python-utilities==3.7.3
+hdx-python-utilities==3.7.4
     # via
     #   hdx-python-api (pyproject.toml)
     #   hdx-python-country
-humanize==4.10.0
+humanize==4.11.0
     # via frictionless
-identify==2.6.0
+identify==2.6.1
     # via pre-commit
-idna==3.8
+idna==3.10
     # via
     #   email-validator
     #   requests
@@ -72,33 +72,33 @@ inflect==7.4.0
     # via quantulum3
 iniconfig==2.0.0
     # via pytest
-isodate==0.6.1
+isodate==0.7.2
     # via frictionless
 jinja2==3.1.4
     # via frictionless
 jsonlines==4.0.0
     # via hdx-python-utilities
-jsonpath-ng==1.6.1
+jsonpath-ng==1.7.0
     # via libhxl
 jsonschema==4.23.0
     # via
     #   frictionless
     #   tableschema-to-template
-jsonschema-specifications==2023.12.1
+jsonschema-specifications==2024.10.1
     # via jsonschema
-libhxl==5.2.1
+libhxl==5.2.2
     # via
     #   hdx-python-api (pyproject.toml)
     #   hdx-python-country
 loguru==0.7.2
     # via hdx-python-utilities
-makefun==1.15.4
+makefun==1.15.6
     # via hdx-python-api (pyproject.toml)
 markdown-it-py==3.0.0
     # via rich
 marko==2.1.2
     # via frictionless
-markupsafe==2.1.5
+markupsafe==3.0.2
     # via jinja2
 mdurl==0.1.2
     # via markdown-it-py
@@ -116,7 +116,7 @@ packaging==24.1
     # via pytest
 petl==1.7.15
     # via frictionless
-platformdirs==4.3.2
+platformdirs==4.3.6
     # via virtualenv
 pluggy==1.5.0
     # via pytest
@@ -126,27 +126,27 @@ ply==3.11
     #   libhxl
 pockets==0.9.1
     # via sphinxcontrib-napoleon
-pre-commit==3.8.0
+pre-commit==4.0.1
     # via hdx-python-api (pyproject.toml)
-pyasn1==0.6.0
+pyasn1==0.6.1
     # via
     #   pyasn1-modules
     #   rsa
-pyasn1-modules==0.4.0
+pyasn1-modules==0.4.1
     # via google-auth
-pydantic==2.9.1
+pydantic==2.9.2
     # via frictionless
-pydantic-core==2.23.3
+pydantic-core==2.23.4
     # via pydantic
 pygments==2.18.0
     # via rich
 pyphonetics==0.5.3
     # via hdx-python-country
-pytest==8.3.2
+pytest==8.3.3
     # via
     #   hdx-python-api (pyproject.toml)
     #   pytest-cov
-pytest-cov==5.0.0
+pytest-cov==6.0.0
     # via hdx-python-api (pyproject.toml)
 python-dateutil==2.9.0.post0
     # via
@@ -186,9 +186,9 @@ requests-oauthlib==2.0.0
     # via google-auth-oauthlib
 rfc3986==2.0.0
     # via frictionless
-rich==13.8.0
+rich==13.9.4
     # via typer
-rpds-py==0.20.0
+rpds-py==0.21.0
     # via
     #   jsonschema
     #   referencing
@@ -196,20 +196,19 @@ rsa==4.9
     # via google-auth
 ruamel-yaml==0.18.6
     # via hdx-python-utilities
-ruamel-yaml-clib==0.2.8
+ruamel-yaml-clib==0.2.12
     # via ruamel-yaml
-setuptools==74.1.2
+setuptools==75.3.0
     # via ckanapi
 shellingham==1.5.4
     # via typer
-simpleeval==0.9.13
+simpleeval==1.0.3
     # via frictionless
 simplejson==3.19.3
     # via ckanapi
 six==1.16.0
     # via
     #   ckanapi
-    #   isodate
     #   pockets
     #   python-dateutil
     #   sphinxcontrib-napoleon
@@ -223,11 +222,13 @@ tableschema-to-template==0.0.13
     # via hdx-python-utilities
 tabulate==0.9.0
     # via frictionless
+tenacity==9.0.0
+    # via hdx-python-country
 text-unidecode==1.3
     # via python-slugify
-typeguard==4.3.0
+typeguard==4.4.1
     # via inflect
-typer==0.12.5
+typer==0.13.0
     # via frictionless
 typing-extensions==4.12.2
     # via
@@ -240,13 +241,13 @@ unidecode==1.3.8
     # via
     #   libhxl
     #   pyphonetics
-urllib3==2.2.2
+urllib3==2.2.3
     # via
     #   libhxl
     #   requests
 validators==0.34.0
     # via frictionless
-virtualenv==20.26.4
+virtualenv==20.27.1
     # via pre-commit
 wheel==0.44.0
     # via libhxl
